app restructuring
