+++ cryptus-app/components/signup/signup.module.scss
@@ -1,0 +1,24 @@
+@use "/styles/variables" as v;
+
+.container {
+  flex: 1 1 auto;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: min(10%, 128px);
+  gap: 32px;
+  font-size: 24px;
+}
+
+.top {
+  flex: 1 1 auto;
+}
+.middle {
+  flex: 1 1 auto;
+  font-weight: 200;
+}
+.bottom {
+  flex: 1 1 auto;
+  font-weight: 200;
+}
