+++ cryptus-app/package.json
@@ -25,6 +25,7 @@
     "react-dom": "17.0.1",
     "react-dropzone": "^11.2.4",
     "react-pdf": "^5.1.0",
+    "sass": "^1.48.0",
     "tailwind": "^4.0.0",
     "web3": "^1.4.0",
     "webpack": "^4.46.0"
