+++ cryptus-app/styles/_variables.scss
@@ -1,0 +1,10 @@
+$mobile: 768px;
+
+$header-h: 64px;
+$footer-h: 128px;
+$title-h: 200px;
+$title-h-sm: 76px;
+
+$bg-color: #fbf8ee;
+$text-color: #292929;
+$header-color: #acacac;
