+++ cryptus-app/tsconfig.json
@@ -1,11 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": false,
@@ -21,9 +17,13 @@
   "include": [
     "next-env.d.ts",
     "**/*.ts",
-    "**/*.tsx"
-, "pages/api/auth/[...nextauth].ts", "pages/index.tsx", "pages/loginpage.tsx", "pages/loginpage2.jsx", "components/UserForm/UserDetails.jsx", "components/UserForm/UserDetails.jsx"  ],
-  "exclude": [
-    "node_modules"
-  ]
+    "**/*.tsx",
+    "pages/api/auth/[...nextauth].ts",
+    "pages/index.tsx",
+    "pages/login.tsx",
+    "pages/loginpage2.jsx",
+    "components/UserForm/UserDetails.jsx",
+    "components/UserForm/UserDetails.jsx"
+  ],
+  "exclude": ["node_modules"]
 }
